Realtime thread registration
You must call AERealtimeWatchdogStartMonitoring and AERealtimeWatchdogStopMonitoring at the beginning and end of any coreAudio realtime thread. All this does is set the static thread pointer to the address of the CoreAudio thread.
It's not as simple as dropping the library into a project and it magically working, like before, but performance is increased by a large amount if you call lots of obj-c (a quick profile run showed about 10% cpu savings).
Uncomment #define USE_WATCHDOG_REGISTRATION_FUNCTIONS 1 to try it out.